[MODIFY]: move methods to check query from ServerParse class to a new extracted class ServerParseCheck to decrease cyclometric complexity of ServerParse class
